refactor : rename files to kebab case
refs: #8
