+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { HeaderComponent } from "./ngxFormDemo/shared/header/header.component";
+import { HeaderComponent } from "./ngx-form-demo/shared/header/header.component";
 
 @Component({
   selector: 'app-root',
