+++ src/app/ngx-form-lib/index.ts
@@ -1,0 +1,1 @@
+export * from './models/ngx-button-config'
