--- src/app/ngxFormLib/index.ts
@@ -1,1 +1,0 @@
-export * from './models/NgxButtonConfig'
